src/aiori_agent/requirements.txt
typer
nats-py[nkeys]
watchdog
fastapi
uvicorn
pydantic
pydantic-settings
netifaces
PyJWT
pydantic